Tests can now be executed
